third_party/WebKit/LayoutTests/fast/scroll-behavior/scroll-into-view-scrolls-layout-viewport-expected.txt

Test that element.scrollIntoView() only scrolls the layout viewport when the intertVisualViewport is set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.visualViewportScrollX() is window.scrollX
PASS internals.visualViewportScrollY() is window.scrollY
PASS window.scrollX is 0
PASS window.scrollY is > 0
PASS internals.visualViewportScrollX() is window.scrollX
PASS internals.visualViewportScrollY() is window.scrollY
PASS window.scrollX is 0
PASS window.scrollY is > 0
PASS successfullyParsed is true

TEST COMPLETE
